# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  template_internal_http_datatransfers_requests.SSONativeRequest:
    properties:
      code:
        type: string
      code_verifier:
        type: string
      redirect_uri:
        description: |-
          RedirectURI нь native client-д бүртгэгдсэн (жишээ
          geregetemp://oauth2/callback) байх ёстой — code exchange-д яг тааруулна.
        type: string
    required:
    - code
    - code_verifier
    - redirect_uri
    type: object
...
  template_internal_http_datatransfers_requests.SuperadminAccessModeRequest:
    properties:
      mode:
        enum:
        - public
        - private
        type: string
    required:
    - mode
    type: object
...
  template_internal_http_handlers_v1.BaseResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Government Developer Portal V3.0
    url: https://github.com/snykk/go-rest-boilerplate
  description: chi (net/http) + pgx (PostgreSQL) + Redis дээр суурилсан Clean Architecture
    бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib
    Fikri)-ээс үүсэлтэй; HTTP давхаргыг chi, өгөгдлийн давхаргыг pgx руу хөрвүүлсэн.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/snykk/go-rest-boilerplate
  title: Government Developer Portal V3.0 API
  version: "1.0"
paths:
...
  /integrations/{provider}:
    delete:
      description: Тухайн хэрэглэгчийн нэг провайдерын хадгалсан токеныг устгана (idempotent).
      parameters:
      - description: Provider id (google-drive, dropbox, google-meet)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Disconnected
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "400":
          description: Unknown provider
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "401":
          description: Unauthenticated
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Интеграцийг салгах
      tags:
      - integrations
...
  /sso/native:
    post:
      consumes:
      - application/json
      description: Mobile app (iOS/Android) ASWebAuthenticationSession-ийн PKCE code-ийг
        public client-ээр (client_secret-гүй, code_verifier-тэй) солин, иргэнийг upsert
        хийж JWT хос олгоно. State шалгалтгүй (PKCE хамгаална). BFF нь token/refresh_token-ийг
        httpOnly cookie-д суулгана.
      parameters:
      - description: Native PKCE code exchange
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.SSONativeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "400":
          description: Missing code/code_verifier
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      summary: dgov SSO native (mobile PKCE) нэвтрэлт
      tags:
      - sso
...
  /v1/superadmin/access-mode:
    get:
      description: Платформын хандалтын горим (public|private)-ыг буцаана. Зөвхөн
        super admin хандана.
      produces:
      - application/json
      responses:
        "200":
          description: Access mode fetched
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "403":
          description: Not a super admin
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Платформын хандалтын горим
      tags:
      - superadmin
    put:
      consumes:
      - application/json
      description: 'Платформын хандалтын горимыг тохируулна. public: хэн ч Government
        SSO-оор нэвтэрнэ; private: зөвхөн урьдчилан бүртгэсэн хэрэглэгч. Зөвхөн super
        admin хандана.'
      parameters:
      - description: Access mode
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/template_internal_http_datatransfers_requests.SuperadminAccessModeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Access mode updated
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
        "403":
          description: Not a super admin
          schema:
            $ref: '#/definitions/template_internal_http_handlers_v1.BaseResponse'
      security:
      - BearerAuth: []
      summary: Платформын хандалтын горим тохируулах
      tags:
      - superadmin
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: /auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен
      (access token).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"